requirements.txt
autopep8==1.5.4
cbor2==5.2.0
certifi==2020.12.5
chardet==3.0.4
Deprecated==1.2.10
idna==2.10
luma.core==2.2.0
luma.led-matrix==1.5.0
Pillow==8.0.1
pycodestyle==2.6.0
pycricbuzz==2.4
pyftdi==0.52.0
pyserial==3.5
pyusb==1.1.0
requests==2.25.0
smbus2==0.4.0
toml==0.10.2
urllib3==1.26.2
wrapt==1.12.1
